Fix selected project outdated height when window resized
Whenever the window resizes, the selected project's details height
is updated to the inner content height. This fixes a bug where
the selected project height was either too low,
hiding the buttons, or too high, showing empty space,
depending on the resize.

Also fixed a bug where an in-progress project resize animation would
height when the window was resized during the animation.